Refactor GroundSegmentation as templated filter
- Inherit pcl::Filter
- Template on point type
- Instead of copying points, work with indices
- Don't use PointXYZGD (seemed ultimately unused anyway)
- Add setKeepGround(), setKeepObstacle() and setKeepOverhanging() methods
  for choosing which parts are wanted. By default just ground is removed.
