+++ wave_matching/tests/ground_segmentation_test.cpp
@@ -11,24 +11,17 @@ const auto TEST_SCAN = "tests/data/testscan.pcd";
 const auto TEST_CONFIG = "tests/config/ground_segmentation.yaml";
 
 TEST(ground_segmentation, how_to_use) {
-    GroundSegmentationParams params(TEST_CONFIG);
-    GroundSegmentation ground_segmentation(params);
+    // Prepare the input
+    // @todo factor out pre-filtering steps
 
-    PCLPointCloudPtr ref, vizobs, vizground, vizdrv;
+    PCLPointCloudPtr input =
+      boost::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
+    auto vizground = input->makeShared();
+    auto vizobs = input->makeShared();
+    auto vizdrv = input->makeShared();
 
-    ref = boost::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    vizobs = boost::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    vizground = boost::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    vizdrv = boost::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
 
-    pcl::PointCloud<PointXYZGD>::Ptr groundCloud =
-      boost::make_shared<pcl::PointCloud<PointXYZGD>>();
-    pcl::PointCloud<PointXYZGD>::Ptr obsCloud =
-      boost::make_shared<pcl::PointCloud<PointXYZGD>>();
-    pcl::PointCloud<PointXYZGD>::Ptr drvCloud =
-      boost::make_shared<pcl::PointCloud<PointXYZGD>>();
-
-    pcl::io::loadPCDFile(TEST_SCAN, *ref);
+    pcl::io::loadPCDFile(TEST_SCAN, *input);
 
     // filter points on the car
     pcl::ConditionalRemoval<pcl::PointXYZ> condrem;
@@ -45,7 +38,6 @@ TEST(ground_segmentation, how_to_use) {
         "x", pcl::ComparisonOps::LT, -3)));
     x_cond->addComparison(pcl::FieldComparison<pcl::PointXYZ>::ConstPtr(
       new pcl::FieldComparison<pcl::PointXYZ>("x", pcl::ComparisonOps::GT, 3)));
-
     y_cond->addComparison(pcl::FieldComparison<pcl::PointXYZ>::ConstPtr(
       new pcl::FieldComparison<pcl::PointXYZ>(
         "y", pcl::ComparisonOps::LT, -1.1)));
@@ -58,16 +50,26 @@ TEST(ground_segmentation, how_to_use) {
 
     condrem.setCondition(total_cond);
 
-    condrem.setInputCloud(ref);
-    condrem.filter(*ref);
+    condrem.setInputCloud(input);
+    condrem.filter(*input);
 
-    ground_segmentation.setupGroundSegmentation(
-      ref, groundCloud, obsCloud, drvCloud);
-    ground_segmentation.segmentGround();
+    // Test the actual ground segmentation filter
+    GroundSegmentationParams params{TEST_CONFIG};
+    GroundSegmentation<pcl::PointXYZ> ground_segmentation{params};
+    ground_segmentation.setInputCloud(input);
 
-    pcl::copyPointCloud(*groundCloud, *vizground);
-    pcl::copyPointCloud(*obsCloud, *vizobs);
-    pcl::copyPointCloud(*drvCloud, *vizdrv);
+    ground_segmentation.setKeepGround(true);
+    ground_segmentation.setKeepObstacle(false);
+    ground_segmentation.setKeepOverhanging(false);
+    ground_segmentation.filter(*vizground);
+
+    ground_segmentation.setKeepGround(false);
+    ground_segmentation.setKeepObstacle(true);
+    ground_segmentation.filter(*vizobs);
+
+    ground_segmentation.setKeepObstacle(false);
+    ground_segmentation.setKeepOverhanging(true);
+    ground_segmentation.filter(*vizdrv);
 
     PointCloudDisplay display("groundsegmenter");
     display.startSpin();
